Updated git info and telemetryCapture updates

+++ telemetryCapture.py
@@ -3,60 +3,36 @@ from struct import unpack
 import data_packet
 from datetime import datetime
 import csv
+import os
 
-'''Class is meant to be used as the main telemetry capture method. It is required to have launched the game before starting this script. Once the game is launched and default variables within the script have been updated (if required) you can launch this script. Current version has no built in kill switch other than keyboard interrupt within the terminal.'''
+'''Class is meant to be used as the main telemetry capture method. It is required to have launched the game before starting this script or it may error out. Once the game is launched and default variables within the script have been updated (if required) you can launch this script. Current version has no built in kill switch other than keyboard interrupt within the terminal.'''
 
-'''Method to handle creation of the CSV file and fill the header with the supplied values'''
+#Method to handle creation of the CSV file and fill the header with the supplied values
 def createHeaderCSV(fileName):
     '''Sled values currently not in scope that could be added back
-        'active', 
-        'acceleration',
-        'velocity',
-        'angular_velocity',
-        'yaw', 'pitch', 'roll',
-        'suspension_travel_ratio',
-        'wheel_slip_ratio',
-        'wheel_rotation_speed',
-        'wheel_on_rumble_strip',
-        'wheel_puddle_depth',
-        'surface_rumble',
-        'wheel_slip_angle',
-        'wheel_combined_slip',
-        'suspension_travel',
+        'active', 'acceleration', 'velocity', 'angular_velocity', 'yaw', 'pitch', 'roll','suspension_travel_ratio','wheel_slip_ratio','wheel_rotation_speed','wheel_on_rumble_strip','wheel_puddle_depth','surface_rumble','wheel_slip_angle','wheel_combined_slip','suspension_travel',
     Dash values currently not in scope that could be added back
-       'position'
-       'tire_temp'
-       'dist_traveled',
-       'lap_time_best', 
-       'lap_time_last', 
-       'lap_time_current',
-       'race_time', 
-       'lap_num', 
-       'race_position',
-       'steering_angle',
-       'driving_line', 
-       'ai_brake_diff'
-       'tire_wear_front_left', 
-       'tire_wear_front_right', 
-       'tire_wear_rear_left', 
-       'tire_wear_rear_right',
-       'track_ordinal_id' '''
+       'position''tire_temp''dist_traveled','lap_time_best', 'lap_time_last', 'lap_time_current','race_time', 'lap_num', 'race_position','steering_angle','driving_line', 'ai_brake_diff''tire_wear_front_left', 'tire_wear_front_right','tire_wear_rear_left', 'tire_wear_rear_right','track_ordinal_id' '''
     
     #Defines values to be placed in the header
     headerData = ["Timestamp", "CarOrdinalID", "CarClass", "CarPerformanceIndex", "DriveTrainType", "NumCylinders", "EngineMaxRPM", "EndingeIdelRPM", "EngineCurrentRPM", "Speed", "Power", "Torque", "Boost", "Gear", "Accel", "Brake", "Clutch", "HandBrake", "Fuel"]
     
     #Writing to the CSV file
-    with open(fileName, mode='w', newline='') as file:
-        writer = csv.writer(file)
-        
-        #Write the header
-        writer.writerow(headerData)
+    if not os.path.exists(fileName):
+        with open(fileName, mode='w', newline='') as file:
+            writer = csv.writer(file)
+            
+            #Write the header
+            writer.writerow(headerData)
 
-        file.close
+            file.close
+        print(f"CSV file '{fileName}' created successfully.")
+    else:
+        print(f"CSV file '{fileName}' already exists.")
 
 #Set default variables to be changed if needed
 gameVersion = "dash_fm8"
-fileName = "logTelemetry.csv"
+fileName = "logTelemetry2.csv"
 
 #Sets UDP information for Forza Motorsport connection
 UDP_IP = "127.0.0.1"
